requirements.txt
crewai==0.76.9
crewai-tools==0.13.4
pytrends==4.9.2
praw==7.8.1
langchain==0.3.7
python-dotenv
pandas